🎨 Simplifies contact flow
