--- app/contact/form/page.tsx
@@ -1,9 +1,0 @@
-import { ContactForm } from "@/components/contact-form";
-
-export default function ContactFormPage() {
-  return (
-    <div>
-      <ContactForm />
-    </div>
-  );
-}
