+++ app/thank-you/page.tsx
@@ -1,0 +1,20 @@
+import { Metadata } from "next";
+import Link from "next/link";
+
+import { SimpleLayout } from "@/components/simple-layout";
+import { buttonVariants } from "@/components/ui/button";
+import { PAGE_METADATA } from "@/content/pages";
+
+export const metadata: Metadata = { ...PAGE_METADATA["thank-you"] };
+export default function ThankYouPage() {
+  return (
+    <SimpleLayout
+      title="Thanks for getting in touch."
+      intro="I'm looking forward to learning more about your project and I'll get back to you on the contact details provided as soon as I can. You can also rest assured that your details are kept safe and not passed onto anyone else without your express permission"
+    >
+      <Link className={buttonVariants({ variant: "default" })} href="/">
+        Back home
+      </Link>
+    </SimpleLayout>
+  );
+}
